+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
 ## Version v1.1.0
-- Fix path issues
 - Changed default backup disk from `s3-backup` to `backup`
+- Fix path issues
+- Fix local only logic